+++ postcss.config.cjs
@@ -2,17 +2,16 @@ module.exports = {
   plugins: [
     require("tailwindcss"),
     require("autoprefixer"),
+    // Prefixing all
     require("postcss-prefix-selector")({
       prefix: ".dev-error-boundary",
-      transform: function (prefix, selector, prefixedSelector) {
-        if (selector === "html") {
-          return prefix;
-        }
+      transform(prefix, selector, prefixedSelector) {
+        if (selector.includes("dev-error-boundary")) return selector;
 
-        if (selector === "body") {
-          return prefix + " " + selector;
+        // Handle special cases for global selectors like body, html, etc.
+        if (/^(body|html|:root)/.test(selector)) {
+          return selector.replace(/^(body|html|:root)/, `$1 ${prefix}`);
         }
-
         return prefixedSelector;
       },
     }),