+++ src/react/react.tsx
@@ -1,1 +1,2 @@
-export { withErrorBoundary } from "./withErrorBoundary";
+export { withErrorBoundary } from "./wrappers/withErrorBoundary";
+export { withErrorBoundaryLinks } from "./wrappers/withErrorBoundaryLinks";