+++ tailwind.config.js
@@ -1,3 +1,8 @@
+import {
+  scopedPreflightStyles,
+  isolateInsideOfContainer, // there are also isolateOutsideOfContainer and isolateForComponents
+} from "tailwindcss-scoped-preflight";
+
 const REM_FACTOR = 0.25;
 
 const spacing = Array.from(Array(1001).keys()).reduce(
@@ -8,12 +13,15 @@ const spacing = Array.from(Array(1001).keys()).reduce(
 /** @type {import('tailwindcss').Config} */
 export default {
   content: ["./src/**/*.tsx", "./src/**/*.ts"],
-  prefix: "mt-",
+  important: true,
+  corePlugins: {
+    preflight: false,
+  },
   theme: {
     extend: {
       maxWidth: ({ theme }) => ({ ...theme("spacing"), "8xl": "90rem" }),
       spacing,
     },
   },
-  plugins: [],
+  plugins: [require("@tailwindcss/forms")],
 };