+++ CHANGELOG.md
@@ -1,5 +1,89 @@
 # @metronome-sh/dev-error-boundary
 
+## 0.0.3
+
+### Patch Changes
+
+- fixed React bug message
+
+## 0.0.2
+
+### Patch Changes
+
+- aac4249: fixed fouc
+- 897699f: found a better way to apply styles
+- Fixed css + React error boundary
+- aac4249: now the error is full screen
+- 60f2947: Fixes
+- 60f2947: removed console.logs
+- 897699f: scoped tailwindcss
+- more css fixes
+- 97493d4: fixed more typings
+- 60f2947: fixed styles import
+- 97493d4: prerelease
+- aac4249: fixed css scoping problems
+- 97493d4: fixed types
+
+## 0.0.2-prerelease.12
+
+### Patch Changes
+
+- more css fixes
+
+## 0.0.2-prerelease.11
+
+### Patch Changes
+
+- Fixed css + React error boundary
+
+## 0.0.2-prerelease.10
+
+### Patch Changes
+
+- scoped tailwindcss
+
+## 0.0.2-prerelease.9
+
+### Patch Changes
+
+- found a better way to apply styles
+
+## 0.0.2-prerelease.8
+
+### Patch Changes
+
+- now the error is full screen
+
+## 0.0.2-prerelease.7
+
+### Patch Changes
+
+- fixed fouc
+
+## 0.0.2-prerelease.6
+
+### Patch Changes
+
+- fixed css scoping problems
+
+## 0.0.2-prerelease.5
+
+### Patch Changes
+
+- fixed styles import
+
+## 0.0.2-prerelease.4
+
+### Patch Changes
+
+- removed console.logs
+
+## 0.0.2-prerelease.3
+
+### Patch Changes
+
+- Fixes
+
 ## 0.0.2-prerelease.2
 
 ### Patch Changes